Improved SearchableList
