+++ src/utils/activityUtils.js
@@ -2,7 +2,7 @@ import { v4 as uuidv4 } from 'uuid';
 import { proper } from './stringUtils';
 import activityLookup from '../store/activityLookup'
 
-const cleanUp = string => string.trim().toLowerCase();
+const cleanUp = string => string.toLowerCase();
 
 //Only handles the iteration, does not know how the matching is done.
 const findActivity = (activityName, matcher) => {
@@ -19,6 +19,11 @@ const exactMatch = activityName =>
 const approximateMatch = activityName =>
     findActivity(activityName, (name, activityName) => name.startsWith(activityName));
 
+const transformInput = (search, activity) => ({
+    searchTerm: search.toLowerCase(),
+    activity: activity.toLowerCase()
+});
+
 //Decides what matcher to use
 export const getActivity = activityName => activityName.includes(":") ?
     approximateMatch(activityName) :
@@ -26,9 +31,19 @@ export const getActivity = activityName => activityName.includes(":") ?
 
 export const findActivityProperty = (activityName, property) => getActivity(activityName)?.[property] ?? null;
 
-export const createActivity = name => ({
-    id: uuidv4(),
-    name: proper(name),
-    category: findActivityProperty(name, "category"),
-    avgTime: findActivityProperty(name, "avgTime"),
-});
+export const createActivity = name => {
+    const activity = {
+        id: uuidv4(),
+        name: proper(name),
+        category: findActivityProperty(name, "category"),
+        avgTime: findActivityProperty(name, "avgTime"),
+    };
+    return activity;
+};
+
+
+export const filterActivities = (activities, searchText) =>
+    activities.filter(item => {
+        const { searchTerm, activity } = transformInput(searchText, item.name);
+        return searchTerm && activity.startsWith(searchTerm) && activity !== searchTerm;
+    });
